requirements.txt
pyalsaaudio
bottle
pyaudio
soundmeter